product-service/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
requests
certifi